Protecting Delete User/Product routes if access by non admin

+++ server/api/products.js
@@ -31,6 +31,11 @@ router.put('/:id', async (req, res, next) => {
 })
 
 router.delete('/:id', async (req, res, next) => {
+  if (!req.user.isAdmin) {
+    const err = new Error('Not allowed!')
+    err.status = 401
+    return next(err)
+  }
   try {
     await Product.destroy({
       where: {
